+++ coil-base/src/androidTest/java/coil/EventListenerTest.kt
@@ -137,9 +137,6 @@ class EventListenerTest {
     @Test
     fun error() {
         val eventListener = TestEventListener(
-            onStart = MethodChecker(false),
-            resolveSizeStart = MethodChecker(false),
-            resolveSizeEnd = MethodChecker(false),
             fetchStart = MethodChecker(false),
             fetchEnd = MethodChecker(false),
             decodeStart = MethodChecker(false),
@@ -160,6 +157,41 @@ class EventListenerTest {
             try {
                 imageLoader.testEnqueue {
                     data("fake_data")
+                }
+            } catch (_: Exception) {}
+        }
+
+        eventListener.complete()
+    }
+
+    @Test
+    fun nullData() {
+        val eventListener = TestEventListener(
+            onStart = MethodChecker(false),
+            resolveSizeStart = MethodChecker(false),
+            resolveSizeEnd = MethodChecker(false),
+            mapStart = MethodChecker(false),
+            mapEnd = MethodChecker(false),
+            fetchStart = MethodChecker(false),
+            fetchEnd = MethodChecker(false),
+            decodeStart = MethodChecker(false),
+            decodeEnd = MethodChecker(false),
+            transformStart = MethodChecker(false),
+            transformEnd = MethodChecker(false),
+            transitionStart = MethodChecker(false),
+            transitionEnd = MethodChecker(false),
+            onSuccess = MethodChecker(false),
+            onCancel = MethodChecker(false)
+        )
+
+        val imageLoader = ImageLoader.Builder(context)
+            .eventListener(eventListener)
+            .build()
+
+        runBlocking {
+            try {
+                imageLoader.testEnqueue {
+                    data(null)
                 }
             } catch (_: Exception) {}
         }
@@ -204,11 +236,11 @@ class EventListenerTest {
     }
 
     private class TestEventListener(
-        val mapStart: MethodChecker = MethodChecker(true),
-        val mapEnd: MethodChecker = MethodChecker(true),
         val onStart: MethodChecker = MethodChecker(true),
         val resolveSizeStart: MethodChecker = MethodChecker(true),
         val resolveSizeEnd: MethodChecker = MethodChecker(true),
+        val mapStart: MethodChecker = MethodChecker(true),
+        val mapEnd: MethodChecker = MethodChecker(true),
         val fetchStart: MethodChecker = MethodChecker(true),
         val fetchEnd: MethodChecker = MethodChecker(true),
         val decodeStart: MethodChecker = MethodChecker(true),
@@ -222,11 +254,11 @@ class EventListenerTest {
         val onError: MethodChecker = MethodChecker(true)
     ) : EventListener {
 
-        override fun mapStart(request: ImageRequest, input: Any) = mapStart.call()
-        override fun mapEnd(request: ImageRequest, output: Any) = mapEnd.call()
         override fun onStart(request: ImageRequest) = onStart.call()
         override fun resolveSizeStart(request: ImageRequest, sizeResolver: SizeResolver) = resolveSizeStart.call()
         override fun resolveSizeEnd(request: ImageRequest, sizeResolver: SizeResolver, size: Size) = resolveSizeEnd.call()
+        override fun mapStart(request: ImageRequest, input: Any) = mapStart.call()
+        override fun mapEnd(request: ImageRequest, output: Any) = mapEnd.call()
         override fun fetchStart(request: ImageRequest, fetcher: Fetcher<*>, options: Options) = fetchStart.call()
         override fun fetchEnd(request: ImageRequest, fetcher: Fetcher<*>, options: Options, result: FetchResult) = fetchEnd.call()
         override fun decodeStart(request: ImageRequest, decoder: Decoder, options: Options) = decodeStart.call()
@@ -240,11 +272,11 @@ class EventListenerTest {
         override fun onError(request: ImageRequest, throwable: Throwable) = onError.call()
 
         fun complete() {
-            mapStart.complete("mapStart")
-            mapEnd.complete("mapEnd")
             onStart.complete("onStart")
             resolveSizeStart.complete("resolveSizeStart")
             resolveSizeEnd.complete("resolveSizeEnd")
+            mapStart.complete("mapStart")
+            mapEnd.complete("mapEnd")
             fetchStart.complete("fetchStart")
             fetchEnd.complete("fetchEnd")
             decodeStart.complete("decodeStart")
