+++ coil-base/src/androidTest/java/coil/fetch/FileFetcherTest.kt
@@ -27,7 +27,7 @@ class FileFetcherTest {
 
     @Test
     fun basic() {
-        val fetcher = FileFetcher(true)
+        val fetcher = FileFetcher()
         val file = context.copyAssetToFile("normal.jpg")
 
         assertTrue(fetcher.handles(file))
@@ -44,7 +44,7 @@ class FileFetcherTest {
 
     @Test
     fun fileCacheKeyWithLastModified() {
-        val fetcher = FileFetcher(true)
+        val fetcher = FileFetcher()
         val file = context.copyAssetToFile("normal.jpg")
 
         file.setLastModified(1234L)
@@ -54,19 +54,5 @@ class FileFetcherTest {
         val secondKey = fetcher.key(file)
 
         assertNotEquals(secondKey, firstKey)
-    }
-
-    @Test
-    fun fileCacheKeyWithoutLastModified() {
-        val fetcher = FileFetcher(false)
-        val file = context.copyAssetToFile("normal.jpg")
-
-        file.setLastModified(1234L)
-        val firstKey = fetcher.key(file)
-
-        file.setLastModified(4321L)
-        val secondKey = fetcher.key(file)
-
-        assertEquals(secondKey, firstKey)
     }
 }
